backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMySQL==1.1.0
requests==2.31.0
numpy>=1.26.0
scikit-learn>=1.3.0
python-dotenv==1.0.0
